Extracted functions and states fro m App.js to NewEventForm.js
